+++ src/app/static/src/app/store/plottingSelections/plottingSelections.ts
@@ -1,5 +1,4 @@
 import {FilterOption} from "../../generated";
-import {localStorageManager} from "../../localStorageManager";
 import {Module} from "vuex";
 import {mutations} from "./mutations";
 import {getters} from "./getters";
@@ -138,13 +137,14 @@ export const initialPlottingSelectionsState = (): PlottingSelectionsState => {
 };
 
 const namespaced = true;
-const existingState = localStorageManager.getState();
 
-export const plottingSelections: Module<PlottingSelectionsState, DataExplorationState> = {
-    namespaced,
-    state: {...initialPlottingSelectionsState(), ...existingState && existingState.plottingSelections},
-    mutations,
-    getters
+export const plottingSelections = (existingState: Partial<DataExplorationState> | null): Module<PlottingSelectionsState, DataExplorationState> => {
+    return {
+        namespaced,
+        state: {...initialPlottingSelectionsState(), ...existingState && existingState.plottingSelections},
+        mutations,
+        getters
+    }
 };