+++ src/app/static/src/tests/components/app.test.ts
@@ -136,7 +136,7 @@ describe("App", () => {
 
         expect(spy).toHaveBeenCalled();
         expect(spy.mock.calls[0][0]?.language).toBe("pt");
-        expect(localStorageManager.getState()?.language).toEqual("pt")
+        expect(localStorageManager.getState(false)?.language).toEqual("pt")
     });
 
     it("resets inputs if baseline update mutation is called and state is ready", () => {