+++ src/app/static/src/tests/components/dataExploration/dataExploration.test.ts
@@ -3,10 +3,11 @@ import DataExploration from "../../../app/components/dataExploration/DataExplora
 import Vuex from "vuex";
 import {emptyState} from "../../../app/root";
 import {
-    mockBaselineState,
-    mockSurveyAndProgramState
+    mockBaselineState, mockProgramResponse, mockShapeResponse,
+    mockSurveyAndProgramState, mockSurveyResponse
 } from "../../mocks";
 import {getters} from "../../../app/store/surveyAndProgram/getters";
+import {baselineGetters} from "../../../app/store/baseline/baseline";
 import registerTranslations from "../../../app/store/translations/registerTranslations";
 import {BaselineActions} from "../../../app/store/baseline/actions";
 import {SurveyAndProgramActions} from "../../../app/store/surveyAndProgram/actions";
@@ -20,8 +21,9 @@ describe(`data exploration component`, () => {
     let sapActions: jest.Mocked<SurveyAndProgramActions>;
     let sapMutations = {};
 
-    const createSut = (baselineState?: Partial<BaselineState>,
-                       surveyAndProgramState: Partial<SurveyAndProgramState> = {selectedDataType: DataType.Survey}) => {
+    const createStore = (baselineState?: Partial<BaselineState>,
+                         surveyAndProgramState: Partial<SurveyAndProgramState> = {selectedDataType: DataType.Survey},
+                         plottingMetadataMock = jest.fn()) => {
 
         actions = {
             refreshDatasetMetadata: jest.fn(),
@@ -46,7 +48,8 @@ describe(`data exploration component`, () => {
                     namespaced: true,
                     state: mockBaselineState(baselineState),
                     actions: {...actions},
-                    mutations: {...mutations}
+                    mutations: {...mutations},
+                    getters: {...baselineGetters}
                 },
                 surveyAndProgram: {
                     namespaced: true,
@@ -54,6 +57,12 @@ describe(`data exploration component`, () => {
                     mutations: {...sapMutations},
                     actions: {...sapActions},
                     getters: getters
+                },
+                metadata: {
+                    namespaced: true,
+                    actions: {
+                        getPlottingMetadata: plottingMetadataMock
+                    }
                 }
             }
         });
@@ -63,7 +72,7 @@ describe(`data exploration component`, () => {
     };
 
     it(`renders components as expect`, () => {
-        const store = createSut()
+        const store = createStore()
         const wrapper = shallowMount(DataExploration, {store});
         expect(wrapper.find("adr-integration-stub").exists()).toBe(true)
         expect(wrapper.find("upload-inputs-stub").exists()).toBe(true)
@@ -71,14 +80,32 @@ describe(`data exploration component`, () => {
     })
 
     it(`disables back navigation when on upload step`, () => {
-        const store = createSut()
+        const store = createStore()
         const wrapper = shallowMount(DataExploration, {store});
         expect(wrapper.find("stepper-navigation-stub").props("backDisabled")).toBe(true)
+    })
+
+    it(`disables forward navigation when inputs are not valid`, () => {
+        const store = createStore()
+        const wrapper = shallowMount(DataExploration, {store});
+        expect(wrapper.find("stepper-navigation-stub").props("nextDisabled")).toBe(true)
+    })
+
+    it(`enables forward navigation when inputs are valid`, () => {
+        const store = createStore(
+            {
+                shape: mockShapeResponse(),
+                validatedConsistent: true
+            },
+            {
+                program: mockProgramResponse()
+            })
+        const wrapper = shallowMount(DataExploration, {store});
         expect(wrapper.find("stepper-navigation-stub").props("nextDisabled")).toBe(false)
     })
 
     it(`disables continue navigation when on review step`, () => {
-        const store = createSut()
+        const store = createStore({shape: mockShapeResponse()}, {survey: mockSurveyResponse()})
         const wrapper = shallowMount(DataExploration, {
             store,
             data() {
@@ -90,8 +117,8 @@ describe(`data exploration component`, () => {
         expect(wrapper.find("stepper-navigation-stub").props("nextDisabled")).toBe(true)
     })
 
-    it(`continue button can navigate to reviewInputs`, () => {
-        const store = createSut()
+    it(`can navigate to reviewInputs`, () => {
+        const store = createStore()
         const wrapper = shallowMount(DataExploration, {store});
         expect(wrapper.find("review-inputs-stub").exists()).toBe(false)
         const vm = wrapper.vm as any
@@ -99,8 +126,8 @@ describe(`data exploration component`, () => {
         expect(wrapper.find("review-inputs-stub").exists()).toBe(true)
     })
 
-    it(`back button can navigate to uploadInputs`, () => {
-        const store = createSut()
+    it(`can navigate to uploadInputs`, () => {
+        const store = createStore()
         const wrapper = shallowMount(DataExploration, {
             store,
             data() {
@@ -112,6 +139,13 @@ describe(`data exploration component`, () => {
         const vm = wrapper.vm as any
         vm.back()
         expect(wrapper.find("upload-inputs-stub").exists()).toBe(true)
-    })
+    });
+
+    it("requests plotting metadata on mount", () => {
+        const metadataMock = jest.fn();
+        const store = createStore({}, {}, metadataMock)
+        const wrapper = shallowMount(DataExploration, {store})
+        expect(metadataMock.mock.calls.length).toBe(1);
+    });
 
 });