+++ NEWS.md
@@ -1,6 +1,14 @@
-# hint 1.85.0
+# hint 1.87.0
 
 * Formatting Input Time Series table values
+
+# hint 1.86.0
+
+* Persist Data Exploration state
+
+# hint 1.85.0
+
+* Input file validation in Data Exploration
 
 # hint 1.84.0